--- a/benchmark/index.js
+++ b/benchmark/index.js
@@ -5,8 +5,7 @@ var Suite = require('benchmarked');
 var suite = new Suite({
   result: false,
   fixtures: 'fixtures/{5,50,250,2000,20000}.js',
-  add: 'code/{repeat-string,repeating,native}.js',
-  cwd: __dirname
+  code: 'code/{repeat-string,repeat-string-1.5.4,repeating,native}.js',
 });
 
 suite.run();