@jonschlinkert 2016-05-13
wow this great! I might have to merge tomorrow, but I'll do it sooner if possible. thanks!